requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl
flask
twitchio